requirements.txt
numpy
tensorflow
pandas
keras_VGGFace
scikit-image
pillow
matplotlib
opencv-python
